docker/requirements.txt
flask
cassandra-driver
python-dotenv